requirements.txt
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
python-dotenv==1.0.0
